[MM-11280] Migrate activity_log_modal.jsx to use Redux (#1584)
* [MM-11280] Migrate activity_log_modal.jsx to Redux

* [MM-11280] Refactor to smaller components

* [MM-11280] Move logic around to simplify components

* [MM-11280] Update mattermost-redux version
