+++ components/activity_log_modal/index.js
@@ -4,6 +4,7 @@
 import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
 import {getSessions, revokeSession} from 'mattermost-redux/actions/users';
+import {getCurrentUserId, getUserSessions} from 'mattermost-redux/selectors/entities/users';
 
 import {getCurrentLocale} from 'selectors/i18n';
 
@@ -11,6 +12,8 @@ import ActivityLogModal from './activity_log_modal.jsx';
 
 function mapStateToProps(state) {
     return {
+        currentUserId: getCurrentUserId(state),
+        sessions: getUserSessions(state),
         locale: getCurrentLocale(state),
     };
 }
